@@ -1,63 +1,71 @@
-import React from 'react';
+import React, { useState } from 'react'
 
-// class Categories extends React.Component {
-//   state = {
-//     activeItem: 3,
-//     test: 123,
-//   };
+export default function Categories({ items }) {
+    const [activeItem, setActiveItem] = useState(0);
 
-//   onSelectItem = (index) => {
-//     this.setState({
-//       activeItem: index,
-//     });
-//   };
+    const onSelectItem = index => {
+        setActiveItem(index)
+    };
 
-//   render() {
-//     const { items, onClickItem } = this.props;
-//     console.log(this.state);
-//     return (
-//       <div className="categories">
-//         <ul>
-//           <li>Все</li>
-//           {items.map((name, index) => (
-//             <li
-//               className={this.state.activeItem === index ? 'active' : ''}
-//               onClick={() => this.onSelectItem(index)}
-//               key={`${name}_${index}`}>
-//               {name}
-//             </li>
-//           ))}
-//         </ul>
-//       </div>
-//     );
-//   }
-// }
 
-function Categories({ items, onClickItem }) {
-  const [activeItem, setActiveItem] = React.useState(null);
+    const elem = items.map((name, i) => {
+        const listClassName = activeItem === i ? 'active' : 0;
 
-  const onSelectItem = (index) => {
-    setActiveItem(index);
-  };
+        return <li
+            className={listClassName}
+            onClick={() => onSelectItem(i)}
+            key={Math.random() * 10} > {name}</li >
+    })
 
-  return (
-    <div className="categories">
-      <ul>
-        <li className={activeItem === null ? 'active' : ''} onClick={() => onSelectItem(null)}>
-          Все
-        </li>
-        {items &&
-          items.map((name, index) => (
-            <li
-              className={activeItem === index ? 'active' : ''}
-              onClick={() => onSelectItem(index)}
-              key={`${name}_${index}`}>
-              {name}
-            </li>
-          ))}
-      </ul>
-    </div>
-  );
+    return (
+        <div className="categories">
+            <ul>
+                {elem}
+            </ul>
+        </div>
+    )
 }
 
-export default Categories;
+// import React, { Component } from 'react';
+
+// class Categories extends Component {
+
+//     state = {
+//         activeItem: 0
+//     }
+
+//     onSelectItem = (index) => {
+//         this.setState({
+//             activeItem: index
+//         })
+//     }
+
+//     render() {
+//         const { items, outputName } = this.props;
+//         // Мапим количество листов в зависимости от массива пришедшего от просов
+//         const element = items.map((name, i) => {
+//             // Записываем класс в зависимости от стейта
+//             const listClassName = this.state.activeItem === i ? 'active' : null;
+//             return <li
+//                 className={listClassName}
+//                 onClick={() => {
+//                     this.onSelectItem(i)
+//                     outputName(name)
+//                 }}
+//                 key={Math.random() * 10}>{name}
+//             </li>
+//         })
+
+//         return (
+//             <div className="categories">
+//                 <ul>
+//                     {element}
+//                 </ul>
+//             </div>
+//         )
+//     }
+
+// }
+
+
+// export default Categories;
