@@ -1,89 +1,71 @@
 import React from 'react';
-import PropTypes from 'prop-types';
-import classNames from 'classnames';
+import TypeLists from './TypeLists';
+import SizeLists from './SizeLists';
+import { useSelector } from 'react-redux'
+// import PropTypes from 'prop-types';
 
-function PizzaBlock({ name, imageUrl, price, types, sizes }) {
-  const availableTypes = ['тонкое', 'традиционное'];
-  const availableSizes = [26, 30, 40];
+function PizzaBlock() {
+    const { items } = useSelector(({ pizzas }) => {
+        return {
+            items: pizzas.items
+        }
+    });
 
-  const [activeType, setActiveType] = React.useState(types[0]);
-  const [activeSize, setActiveSize] = React.useState(sizes[0]);
+    const pizzaItems = items.map((item) => {
 
-  const onSelectType = (index) => {
-    setActiveType(index);
-  };
+        const { imageUrl, name, types, price, sizes, id } = item;
 
-  const onSelectSize = (index) => {
-    setActiveSize(index);
-  };
 
-  return (
-    <div className="pizza-block">
-      <img className="pizza-block__image" src={imageUrl} alt="Pizza" />
-      <h4 className="pizza-block__title">{name}</h4>
-      <div className="pizza-block__selector">
-        <ul>
-          {availableTypes.map((type, index) => (
-            <li
-              key={type}
-              onClick={() => onSelectType(index)}
-              className={classNames({
-                active: activeType === index,
-                disabled: !types.includes(index),
-              })}>
-              {type}
-            </li>
-          ))}
-        </ul>
-        <ul>
-          {availableSizes.map((size, index) => (
-            <li
-              key={size}
-              onClick={() => onSelectSize(index)}
-              className={classNames({
-                active: activeSize === index,
-                disabled: !sizes.includes(size),
-              })}>
-              {size} см.
-            </li>
-          ))}
-        </ul>
-      </div>
-      <div className="pizza-block__bottom">
-        <div className="pizza-block__price">от {price} ₽</div>
-        <div className="button button--outline button--add">
-          <svg
-            width="12"
-            height="12"
-            viewBox="0 0 12 12"
-            fill="none"
-            xmlns="http://www.w3.org/2000/svg">
-            <path
-              d="M10.8 4.8H7.2V1.2C7.2 0.5373 6.6627 0 6 0C5.3373 0 4.8 0.5373 4.8 1.2V4.8H1.2C0.5373 4.8 0 5.3373 0 6C0 6.6627 0.5373 7.2 1.2 7.2H4.8V10.8C4.8 11.4627 5.3373 12 6 12C6.6627 12 7.2 11.4627 7.2 10.8V7.2H10.8C11.4627 7.2 12 6.6627 12 6C12 5.3373 11.4627 4.8 10.8 4.8Z"
-              fill="white"
-            />
-          </svg>
-          <span>Добавить</span>
-          <i>2</i>
+        return (
+            <div
+                key={id + (Math.random() * 10)}
+                className="pizza-block">
+                <img
+                    className="pizza-block__image"
+                    src={imageUrl}
+                    alt={name}
+                />
+                <h4 className="pizza-block__title">{name}</h4>
+                <div className="pizza-block__selector">
+                    <ul>
+                        <TypeLists types={types} />
+                    </ul>
+                    <ul>
+                        <SizeLists sizes={sizes} />
+                    </ul>
+                </div>
+                <div className="pizza-block__bottom">
+                    <div className="pizza-block__price">{price}<svg className="ruble-svg"><use xlinkHref="#ruble" /></svg>  </div>
+                    <div className="button button--outline button--add">
+                        <svg
+                            width="12"
+                            height="12"
+                            viewBox="0 0 12 12"
+                            fill="none"
+                            xmlns="http://www.w3.org/2000/svg"
+                        >
+                            <path
+                                d="M10.8 4.8H7.2V1.2C7.2 0.5373 6.6627 0 6 0C5.3373 0 4.8 0.5373 4.8 1.2V4.8H1.2C0.5373 4.8 0 5.3373 0 6C0 6.6627 0.5373 7.2 1.2 7.2H4.8V10.8C4.8 11.4627 5.3373 12 6 12C6.6627 12 7.2 11.4627 7.2 10.8V7.2H10.8C11.4627 7.2 12 6.6627 12 6C12 5.3373 11.4627 4.8 10.8 4.8Z"
+                                fill="white"
+                            />
+                        </svg>
+                        <span>Добавить</span>
+                        <i>2</i>
+                    </div>
+                </div>
+            </div>
+        );
+    });
+
+    return (
+        <div className="content__items">
+            {pizzaItems}
         </div>
-      </div>
-    </div>
-  );
+    )
 }
 
-PizzaBlock.propTypes = {
-  name: PropTypes.string,
-  imageUrl: PropTypes.string,
-  price: PropTypes.number,
-  types: PropTypes.arrayOf(PropTypes.number),
-  sizes: PropTypes.arrayOf(PropTypes.number),
-};
-
-PizzaBlock.defaultProps = {
-  name: '---',
-  price: 0,
-  types: [],
-  sizes: [],
-};
+// PizzaBlock.propTypes = {
+//     name: PropTypes.string.isRequired,
+// }
 
 export default PizzaBlock;
