@@ -1,5 +1,4 @@
-export { default as Button } from './Button';
-export { default as Header } from './Header';
 export { default as Categories } from './Categories';
-export { default as SortPopup } from './SortPopup';
+export { default as Header } from './Header';
+export { default as Sort } from './Sort';
 export { default as PizzaBlock } from './PizzaBlock';
