@@ -1,11 +1,11 @@
-import React from 'react';
+import React from 'react'
 
 function Cart() {
-  return (
-    <div>
-      <h1>Корзина</h1>
-    </div>
-  );
+    return (
+        <div>
+            <h1>Here is cart</h1>
+        </div>
+    )
 }
 
-export default Cart;
+export default Cart
