@@ -1,30 +1,30 @@
 import React from 'react';
-import { Categories, SortPopup, PizzaBlock } from '../components';
+import { Categories, Sort, PizzaBlock } from '../components';
 
-function Home({ items }) {
-  return (
-    <div className="container">
-      <div className="content__top">
-        <Categories
-          onClickItem={(name) => console.log(name)}
-          items={['Мясные', 'Вегетарианская', 'Гриль', 'Острые', 'Закрытые']}
-        />
-        <SortPopup
-          items={[
-            { name: 'популярности', type: 'popular' },
-            { name: 'цене', type: 'price' },
-            { name: 'алфавит', type: 'alphabet' },
-          ]}
-        />
-      </div>
-      <h2 className="content__title">Все пиццы</h2>
-      <div className="content__items">
-        {items.map((obj) => (
-          <PizzaBlock key={obj.id} {...obj} />
-        ))}
-      </div>
-    </div>
-  );
+function Home({ setShowPopup, showPopup, onSelectPopup }) {
+
+    return (
+        <div className="container">
+            <div className="content__top">
+                <Categories
+                    items={[
+                        'Все',
+                        'Мясные',
+                        'Вегетарианские',
+                        'Гриль',
+                        'Острые',
+                        'Закрытые'
+                    ]} />
+                <Sort
+                    setShowPopup={setShowPopup}
+                    showPopup={showPopup}
+                    onSelectPopup={onSelectPopup} />
+            </div>
+
+            <h2 className="content__title">Все пиццы</h2>
+            <PizzaBlock />
+        </div>
+    )
 }
 
 export default Home;
