@@ -1,4 +1,7 @@
-export const setPizzas = (items) => ({
-  type: 'SET_PIZZAS',
-  payload: items,
+
+const setPizzas = (items) => ({
+    type: 'SET_PIZZAS',
+    payload: items
 });
+
+export default setPizzas;
