@@ -1,11 +1,10 @@
 import { combineReducers } from 'redux';
-
-import filtersReducer from './filters';
-import pizzasReducer from './pizzas';
+import filters from './filters'
+import pizzas from './pizzas'
 
 const rootReducer = combineReducers({
-  filters: filtersReducer,
-  pizzas: pizzasReducer,
+    filters,
+    pizzas
 });
 
 export default rootReducer;
