@@ -1,17 +1,16 @@
 const initialState = {
-  items: [],
-  isLoaded: false,
-};
+    items: []
 
+}
 const pizzas = (state = initialState, action) => {
-  console.log(action, 'reducerPizzas');
-  if (action.type === 'SET_PIZZAS') {
-    return {
-      ...state,
-      items: action.payload,
-    };
-  }
-  return state;
-};
+    if (action.type === 'SET_PIZZAS') {
+        return {
+            ...state,
+            items: action.payload
+        }
+    }
+    return state;
+
+}
 
 export default pizzas;
