@@ -1,12 +1,6 @@
 import { createStore } from 'redux';
+import rootReducer from './reducers'
 
-import rootReducer from './reducers';
-
-const store = createStore(
-  rootReducer,
-  window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__(),
-);
-
-window.store = store;
+const store = createStore(rootReducer);
 
 export default store;
